Creación del boostrap de facilities, cambio de float a double como ty…

+++ src/main/kotlin/ar/edu/unsam/phm/magicnightsback/domain/Location.kt
@@ -1,6 +1,6 @@
 package ar.edu.unsam.phm.magicnightsback.domain
 
 data class Location (
-    val latitud: Float,
-    val longitud: Float
+    val latitud: Double,
+    val longitud: Double
 )
